some clarifications
